Add files via upload
This change contains an R Markdown file with the code I used during the analysis, as well as the necessary tables with the data to reproduce the analysis. If someone wishes to completely rerun my analysis, they need to have R installed, set the working directory to the one containing the other files, and install the packages specified in the code. The rest is just smoothly pressing the run button for each chunk :)